Forward broker requests on rbd-mirror relation
To be able to support mirroring of pools with advanced features
the RBD Mirror charm needs more information about the intent
behind pools requested in a deployment. We solve this by
forwarding all the broker requests in the deployment.

It is up to the consumer of the rbd-mirror relation to filter
the requests and relay the ones eligible for use on a remote
cluster.

Change-Id: I16196053bee93bdc4e5c62f5467d9e786b047b30