Update get_mon_hosts to not specify mon ports
Update get_mon_hosts to avoid specifying the port number for the
MON daemons; we use the default so this is not required, and at
Nautilus the MON daemons run both v1 and v2 messenger ports.

Specifying the port in the ceph.conf file disables the v2
messenger port which is not the desired behaviour on upgrade
or new installation.

Change-Id: I5333cc85f755771733fe09e963d5523fd93c16d2
Closes-Bug: #1917742