Add support for object gateway integration
To add support for object gateways this change adds the
radosgw-user interface.

The focal bundle has one rados gateway application as more
than that is not supported on Octopus. The hersute bundle
has two rados gateway applications but does not contain
the lma applications as telegraf is not supported on
hersute.

Depends-On: Ieff1943b02f490559ccd245f60b744fb76a5d832
Change-Id: I38939b9938a5ba2ed6e3fb489f66a255f7aa8fe4