Update to classic charms to build using charmcraft in CI
This update is to ensure that the Zuul Canonical CI builds the charm
before functional tests and ensure that that artifact is used for the
functional tests.  This is to try to ensure that the charm that gets
landed to the charmhub is the same charm that was tested with.

Change-Id: I17cb0cfe88be075d79d94edcdd2cea845bf78b8f
Co-authored-by: Aurelien Lourot <aurelien.lourot@canonical.com>