Pin tox to < 4.0.0
Tox 4.0.0 was recently released and it has several breaking changes.
We pin to < 4.0.0 here. We are planning to move forward only on the
master charm branches.

Tox is also pinned to < 4.0.0 for stable branches in upstream
openstack-zuul-jobs as well as in zosci-config. However, the
requires= section in the charm's tox.ini file ends up installing
tox>4, wiping out the zuul-pinned tox<4 that was already installed
installed. This patch fixes that.

Additionally, accommodate the following changes by hand from
origin/master:
- add libffi-dev in bindep.txt

Related-Bug: #2002788
Change-Id: Ib216b3b5bab7488d36f774c2b602a393ee0ce630