Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I4c9f607213138adff0af434a417048c9d40af2a4
(cherry picked from commit 66ca6be4ba7caf1ab7ec6a5c5a6dbe9b785819ad)