Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: Ia414ab7d30db9443cf0d9d26071ca48b772f8d1e
(cherry picked from commit 41e3db8134c944fef0c83e5c26880157bcb8f453)