Don't set the default pool as "iscsi"
"iscsi" is the name of the metadata pool by default in the charm and not
to be used for tenant data. The charm can fill out the data pool name
automatically with `def data_pool_name` or can accept a user specified
value. After fixing the issue, the "iscsi" pool will be used only for
hosting gateway.conf as a rados object so reflect the purpose into the
application name for the pool too.

Also, update the command syntax not to print the deprecation warning:
"rbd: --user is deprecated, use --id"

Closes-Bug: #2042419
Related-Bug: #2040274
Change-Id: Iec275190854edcc85915d93db233f444c92fbb17