Use single credential for configuring object storage gw in dashboard.
The radosgw relation provides the same credential mapped to the
hostnames of all the related radosgw application units. This extra
information is not needed. Even when multiple radosgw applications
are present in the same Juju model, a single instance of credentials
is sufficient for dashbaord to present Object GW UI.

Closes-Bug: #1986593
Change-Id: Ib56e0bd4fe588219d6b5728d9a71cf09abdd25de
Signed-off-by: Utkarsh Bhatt <utkarsh.bhatt@canonical.com>