Update bundles to specify source config option
The ceph-mon and ceph-osd charms changed the default source config
option to be bobcat in the master branches. This breaks focal tests
as the focal-bobcat cloud archive is not a valid source repository.
Update the config options for the test bundles to be explicit in the
source of package updates.

Change-Id: Ia535b0500d39409f939f0aa58271866efbcd55fb