Allow static ipv6 addresses & binding check
The charm used to only allow dynamic ipv6 addresses.
Following some testing (see LP issue) I ensured that it works.

The charm wouldn't check if the first local address returned
by get_ipv6_addr was the same one that juju provided for the
public endpoint. This resulted in the charm binding ceph to
some random address while correctly using juju's one to connect
to other mons, which none were listening to.
We log a warning in case this happens, but default to the
previous behavior.
Closes-Bug: #2061836

Change-Id: I63eefba2714ccc3189f5260012b376c531052b86