Fix functional tests
This patchset modifies the test bundle so that instead of microceph,
it uses the ceph-mon and ceph-osd charms. It does this for 2 reasons:

First, to work around a transient bug in microceph where keys that
are handled to clients aren't immediately usable. This can be worked
around by introducing timeouts, but it's shaky and so it was decided
to use a different set of charms.

Secondly, since the ceph-osd charm needs to be a virtual machine and
it supports noble, it allows us to fully test the Linux kernel NVMe-oF
driver to connect to the charm's endpoints.